added main menu
