--- Assets/GameManager.cs
@@ -1,1 +1,0 @@
-﻿
