--- Assets/MapIconScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8df68c0a5835c71438ead7be02213ca4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
