--- Assets/RefHolder.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d0ca9ad728516ca469c68276c728a148
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
