--- Assets/Scripts/DataHolder.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ee77ff8f229278e4a831821e160cbcfe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
