+++ Assets/Scripts/Storagestuff.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b0297df13f2deee47b247b3a99276b62
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
