--- Assets/brickwall.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 917ff6a5daad8ab44928a137ffd06ed4
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
